onboarding_assistant/requirements.txt
streamlit>=1.26
openai>=0.27
faiss-cpu>=1.7
numpy